backend/requirements.txt
python-telegram-bot==21.0
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-dotenv==1.0.0
cryptography==42.0.0
pandas==2.2.0
